requirements.txt
aiogram==3.27.0
sqlalchemy[asyncio]==2.0.49
asyncpg==0.31.0
pydantic==2.12.5
python-dotenv==1.2.2
redis==7.4.0
celery[rabbitmq]==5.5.2
minio==7.2.15
